Add files via upload
This pertains to the additional analysis of the similarity between the ELM- and human-produced trajectories based on the critical velocity and curvature values. One of the possible ways to assert the similarity between the true human-generated trajectories and the ones that originate from ELMs would be to attempt to group the observations of two features, namely,  the critical curvature $\kappa_c$ and critical speed $v_c$, into two clusters that correctly contain the instances of human and ELM power-law features. Visual illustrations are presented of the results of k-means, k-medians, and different outcomes of the Gaussian mixture clustering techniques, none of which is able to detect the two classes/clusters correctly in the ELM-human power-law comparison data. In the particular setting of human- and ELM-generated groups of curves, the fact that none of these methods would allow for an accurate separation between them illustrates the indistinguishability between the critical speeds and curvatures of human- and ELM-associated trajectories.